Fix an issue where data fulfillment is not performed on client-side (#6)
* Fix an issue where data fulfillment is not performed on client-side

The problem surfaced when we switched to `componentDidMount` to fetch
data in order to ensure components are rendered on the server-side. We
expect developers to use `fetchRouteData` to hydrate the application
state before rendering.

We also made the change to prevent an issue where a React warning was
logged because `setState` was called unexpectedly on the server-side.

Although I'm not very proud of it, to patch the issue I went back to
using `componentWillMount` with a Node.js runtime check. That will
update the component's state to fetching on the client-side, but on the
server it will just passthrough and render the component.

Additionally, a couple breaking changes were introduced since I renamed
the options passed to the higher-order component so they made more sense
in the context that they are executed.

I took advantage of the breaking change to sort some things that I
wanted to do as well. Like reordering the parameters passed to callbacks
to make route props more predominant and renaming the function on the
server side to also make more sense in its context.

* Check for browser instead.

I think is less volatile, since there are other JS environments
(e.g. Rhino, Nashorn, etc).